Kode-cli-main/requirements.txt
fastapi
uvicorn
openai
pydantic
httpx
sse-starlette
python-dotenv
pytest
pytest-asyncio
httpx
